requirements.txt
streamlit
openai
langchain
faiss-cpu
PyPDF2
python-dotenv
tiktoken